requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI & AI
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.2
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.8.0
# For GPU support, use: faiss-gpu==1.8.0

# Document Processing
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Production
gunicorn==21.2.0

# Local Embeddings
sentence-transformers==3.3.1
